Split the run file apart
