+++ extension/src/router_channel.rs
@@ -4,7 +4,7 @@ use std::time::SystemTime;
 use std::{pin::Pin, sync::Arc};
 
 use futures::channel::mpsc;
-use futures::{Future, SinkExt};
+use futures::SinkExt;
 use http_body_util::{BodyExt, StreamBody};
 use httpdate::fmt_http_date;
 use hyper::body::Body;
@@ -12,20 +12,14 @@ use hyper::{
   body::{Bytes, Frame, Incoming},
   Request, Uri,
 };
-use hyper_util::rt::{TokioExecutor, TokioIo};
-use lambda_runtime::LambdaEvent;
+use hyper_util::rt::TokioIo;
 use rand::prelude::*;
-use rand::SeedableRng;
 use tokio::io::{AsyncRead, AsyncWrite};
 use tokio::net::TcpStream;
-use tower::Service;
 
-use crate::cert::AcceptAnyServerCert;
+use crate::app_request;
 use crate::counter_drop::DecrementOnDrop;
-use crate::options::Options;
-use crate::ping;
-use crate::time::{self, current_time_millis};
-use crate::{app_request, messages};
+use crate::time;
 
 use tokio_rustls::client::TlsStream;
 
@@ -37,340 +31,6 @@ impl Stream for TlsStream<TcpStream> {}
 impl Stream for TcpStream {}
 
 type Result<T> = std::result::Result<T, Box<dyn std::error::Error + Send + Sync>>;
-
-#[derive(Clone)]
-pub struct LambdaService {
-  healthcheck_url: Uri,
-  async_init: bool,
-  ready_at_init: Arc<AtomicBool>,
-  domain: Uri,
-  compression: bool,
-}
-
-impl LambdaService {
-  pub fn new(options: &Options) -> Self {
-    let schema = "http";
-
-    let healthcheck_url = format!("{}://{}:{}{}", schema, "127.0.0.1", options.port, "/health")
-      .parse()
-      .unwrap();
-
-    let domain = format!("{}://{}:{}", schema, "127.0.0.1", options.port)
-      .parse()
-      .unwrap();
-
-    LambdaService {
-      async_init: options.async_init,
-      ready_at_init: Arc::new(AtomicBool::new(false)),
-      compression: options.compression,
-      domain,
-      healthcheck_url,
-    }
-  }
-
-  //
-  // This is called by the Tower.Service trait impl below
-  //
-  async fn fetch_response(
-    &self,
-    event: LambdaEvent<messages::WaiterRequest>,
-  ) -> std::result::Result<messages::WaiterResponse, lambda_runtime::Error> {
-    // extract some useful info from the request
-    let lambda_id = event.payload.id;
-    let channel_count: u8 = event.payload.number_of_channels;
-    let dispatcher_url = event.payload.dispatcher_url;
-
-    // prepare the response
-    let resp = messages::WaiterResponse {
-      id: lambda_id.to_string(),
-    };
-
-    if event.payload.init_only {
-      log::info!(
-        "LambdaId: {} - Returning from init-only request",
-        lambda_id.clone()
-      );
-      return Ok(resp);
-    }
-
-    // If the sent_time is more than a second old, just return
-    // This is mostly needed locally where requests get stuck in the queue
-    let sent_time = chrono::DateTime::parse_from_rfc3339(&event.payload.sent_time).unwrap();
-    if sent_time.timestamp_millis() < (current_time_millis() - 5000).try_into().unwrap() {
-      log::info!(
-        "LambdaId: {} - Returning from stale request",
-        lambda_id.clone()
-      );
-      return Ok(resp);
-    }
-
-    log::info!(
-      "LambdaId: {}, Timeout: {}s - Invoked",
-      lambda_id,
-      (event.context.deadline - current_time_millis()) / 1000
-    );
-    let mut deadline_ms = event.context.deadline;
-    if (deadline_ms - current_time_millis()) > 15 * 60 * 1000 {
-      log::warn!("Deadline is greater than 15 minutes, trimming to 1 minute");
-      deadline_ms = current_time_millis() + 60 * 1000;
-    }
-    // check if env var is set to force deadline for testing
-    if let Ok(force_deadline_secs) = std::env::var("LAMBDA_DISPATCH_FORCE_DEADLINE") {
-      log::warn!("Forcing deadline to {} seconds", force_deadline_secs);
-      let force_deadline_secs: u64 = force_deadline_secs.parse().unwrap();
-      deadline_ms = current_time_millis() + force_deadline_secs * 1000;
-    }
-
-    // run until we get a GoAway or deadline is about to be reached
-    let mut lambda_request = LambdaRequest::new(
-      self.domain.clone(),
-      self.compression,
-      lambda_id.clone(),
-      channel_count,
-      dispatcher_url.parse().unwrap(),
-      deadline_ms,
-    );
-    lambda_request.start().await?;
-
-    Ok(resp)
-  }
-}
-
-// Tower.Service is the interface required by lambda_runtime::run
-impl Service<LambdaEvent<messages::WaiterRequest>> for LambdaService {
-  type Response = messages::WaiterResponse;
-  type Error = lambda_runtime::Error;
-  type Future =
-    Pin<Box<dyn Future<Output = std::result::Result<Self::Response, Self::Error>> + Send>>;
-
-  fn poll_ready(
-    &mut self,
-    _cx: &mut core::task::Context<'_>,
-  ) -> core::task::Poll<std::result::Result<(), Self::Error>> {
-    core::task::Poll::Ready(Ok(()))
-  }
-
-  fn call(&mut self, event: LambdaEvent<messages::WaiterRequest>) -> Self::Future {
-    let adapter = self.clone();
-    Box::pin(async move { adapter.fetch_response(event).await })
-  }
-}
-
-#[derive(Clone)]
-pub struct LambdaRequest {
-  domain: Uri,
-  compression: bool,
-  lambda_id: String,
-  channel_count: u8,
-  dispatcher_url: Uri,
-  cancel_token: tokio_util::sync::CancellationToken,
-  deadline_ms: u64,
-  goaway_received: Arc<AtomicBool>,
-  last_active: Arc<AtomicU64>,
-  rng: rand::rngs::StdRng,
-  requests_in_flight: Arc<AtomicUsize>,
-  count: Arc<AtomicUsize>,
-}
-
-//
-// LambdaRequest handles connecting back to the router, picking up requests,
-// sending ping requests to the router, and sending the requests to the contained app
-// When an invoke completes this is torn down completely
-//
-
-impl LambdaRequest {
-  pub fn new(
-    domain: Uri,
-    compression: bool,
-    lambda_id: String,
-    channel_count: u8,
-    dispatcher_url: Uri,
-    deadline_ms: u64,
-  ) -> Self {
-    LambdaRequest {
-      count: Arc::new(AtomicUsize::new(0)),
-      domain,
-      compression,
-      lambda_id,
-      channel_count,
-      dispatcher_url,
-      cancel_token: tokio_util::sync::CancellationToken::new(),
-      deadline_ms,
-      goaway_received: Arc::new(AtomicBool::new(false)),
-      last_active: Arc::new(AtomicU64::new(0)),
-      rng: rand::rngs::StdRng::from_entropy(),
-      requests_in_flight: Arc::new(AtomicUsize::new(0)),
-    }
-  }
-
-  /// Executes a task with a specified deadline.
-  ///
-  /// # Parameters
-  ///
-  /// * `deadline_ms`: A timestamp in milliseconds since the Unix epoch
-  ///   representing when the Lambda function needs to finish execution.
-  pub async fn start(&mut self) -> anyhow::Result<()> {
-    let start_time = time::current_time_millis();
-    let scheme = self.dispatcher_url.scheme().unwrap().to_string();
-    let use_https = scheme == "https";
-    let host = self
-      .dispatcher_url
-      .host()
-      .expect("uri has no host")
-      .to_string();
-    let port = self.dispatcher_url.port_u16().unwrap_or(80);
-    let addr = format!("{}:{}", host, port);
-    let dispatcher_authority = self.dispatcher_url.authority().unwrap().clone();
-
-    // Setup the connection to the router
-    let tcp_stream = TcpStream::connect(addr).await?;
-    tcp_stream.set_nodelay(true)?;
-
-    let mut root_cert_store = rustls::RootCertStore::empty();
-    for cert in rustls_native_certs::load_native_certs()? {
-      root_cert_store.add(cert).ok(); // ignore error
-    }
-    let mut config = rustls::ClientConfig::builder()
-      .with_root_certificates(root_cert_store)
-      .with_no_client_auth();
-    // We're going to accept non-validatable certificates
-    config
-      .dangerous()
-      .set_certificate_verifier(Arc::new(AcceptAnyServerCert));
-    // Advertise http2
-    config.alpn_protocols = vec![b"h2".to_vec()];
-    let connector = tokio_rustls::TlsConnector::from(Arc::new(config));
-    let domain_host = self
-      .dispatcher_url
-      .host()
-      .ok_or_else(|| "Host not found")
-      .unwrap();
-    let domain = rustls_pki_types::ServerName::try_from(domain_host)?;
-
-    let stream: Box<dyn Stream + Unpin>;
-    if use_https {
-      let tls_stream = connector.connect(domain.to_owned(), tcp_stream).await?;
-      stream = Box::new(tls_stream);
-    } else {
-      stream = Box::new(tcp_stream);
-    }
-    let io = TokioIo::new(Pin::new(stream));
-
-    // Setup the HTTP2 connection
-    let (sender, conn) = hyper::client::conn::http2::handshake(TokioExecutor::new(), io)
-      .await
-      .unwrap();
-
-    let lambda_id_clone = self.lambda_id.clone();
-    tokio::task::spawn(async move {
-      if let Err(err) = conn.await {
-        log::error!(
-          "LambdaId: {} - Router HTTP2 connection failed: {:?}",
-          lambda_id_clone.clone(),
-          err
-        );
-      }
-    });
-
-    let app_url: Uri = self.domain.clone();
-
-    // Send the ping requests in background
-    let scheme_clone = scheme.clone();
-    let host_clone = host.clone();
-    let ping_task = tokio::task::spawn(ping::send_ping_requests(
-      Arc::clone(&self.last_active),
-      Arc::clone(&self.goaway_received),
-      dispatcher_authority.to_string(),
-      sender.clone(),
-      self.lambda_id.clone(),
-      Arc::clone(&self.count),
-      scheme_clone,
-      host_clone,
-      port,
-      self.deadline_ms,
-      self.cancel_token.clone(),
-      Arc::clone(&self.requests_in_flight),
-    ));
-
-    // Startup the request channels
-    let futures = (0..self.channel_count)
-      .map(|channel_number| {
-        let app_url = app_url.clone();
-        let compression_enabled = self.compression.clone();
-        let last_active = Arc::clone(&self.last_active);
-        let goaway_received = Arc::clone(&self.goaway_received);
-        let dispatcher_authority = dispatcher_authority.clone();
-        let sender = sender.clone();
-        let count = Arc::clone(&self.count);
-        let rng = self.rng.clone();
-        let scheme = scheme.clone();
-        let host = host.clone();
-        let port = port;
-        let lambda_id = self.lambda_id.clone();
-        let requests_in_flight = Arc::clone(&self.requests_in_flight);
-
-        // Create a JoinHandle and implicitly return it to be collected in the vector
-        tokio::spawn(async move {
-          //
-          // TODO: Create a channel object and call the start function on it
-          //
-
-          let mut router_channel = RouterChannel::new(
-            Arc::clone(&count),
-            compression_enabled,
-            lambda_id.clone(),
-            Arc::clone(&goaway_received),
-            Arc::clone(&last_active),
-            rng.clone(),
-            Arc::clone(&requests_in_flight),
-            scheme.clone(),
-            host.clone(),
-            port.clone(),
-            app_url.clone(),
-            channel_number.clone(),
-            sender.clone(),
-            dispatcher_authority.clone(),
-          );
-
-          return router_channel.start().await;
-        })
-      })
-      .collect::<Vec<_>>();
-
-    tokio::select! {
-        result = futures::future::try_join_all(futures) => {
-            match result {
-                Ok(_) => {
-                  // All tasks completed successfully
-                }
-                Err(_) => {
-                  panic!("LambdaId: {} - run - Error in futures::future::try_join_all", self.lambda_id.clone());
-                }
-            }
-        }
-    }
-
-    // Wait for the ping loop to exit
-    self.cancel_token.cancel();
-    ping_task.await?;
-
-    // Print final stats
-    let elapsed = time::current_time_millis() - start_time;
-    let rps = format!(
-      "{:.1}",
-      self.count.load(Ordering::Acquire) as f64 / (elapsed as f64 / 1000.0)
-    );
-    log::info!(
-      "LambdaId: {}, Requests: {}, Elapsed: {} ms, RPS: {} - Returning from run",
-      self.lambda_id,
-      self.count.load(Ordering::Acquire),
-      elapsed,
-      rps
-    );
-
-    Ok(())
-  }
-}
 
 #[derive(Clone)]
 pub struct RouterChannel {
@@ -831,7 +491,6 @@ impl RouterChannel {
 
         bytes_read += chunk_len;
 
-        // TODO: This is where we can gzip
         if let Some(ref mut encoder) = encoder {
           let chunk_data = chunk.as_ref().unwrap().data_ref().unwrap();
           encoder.write_all(chunk_data)?;
